test: Test BPF masquerading for IPv6
Since Cilium supports BPF masquerading for IPv6, we can now test it.

The upgrade/downgrade test needs a small exception because IPv6 BPF
masquerading is still not supported on v1.12, from which we upgrade.
The exception logic comes with an assertion, to ensure we don't forget
to remove the exception once the upgrade is performed from v1.14.

[ Quentin:
    - Bump version number requirement in assert from 1.12 to 1.14,
      switch fixed assert to versioncheck.
    - Drop test updates for test/k8s/datapath_configuration.go, no
      longer relevant after commits b5e0e124226e ("test: Remove
      K8sDatapathConfig some direct routing tests") and 84116274b4b4
      ("test: Remove K8sDatapathConfig some tunneling tests"). Instead,
      enable IPv6 masquerading in the datapath conformance tests. ]

Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>