test/verifier: Fix compilation command
As part of the verifier tests, we first compile the datapath then
attempt to load it. We do that for a set of kernels and datapath
configurations. We compile e.g. bpf_lxc with:

    make -C bpf KERNEL=[kernel] MAX_LXC_OPTIONS=[config]

That method of passing the two environment variables however doesn't
seem to work in the CI for some reason. The datapath config used for
the compilation ends up empty and the default one is used instead.

That means the verifier test has been covering a subset of the expected
datapath configurations. The 4.19 kernel is not affected because it was
kept in the previous K8sVerifier ginkgo test.

This commit fixes it by using cmd.Env instead to pass the environment
variables. Note we also need to switch from %q to %s for the datapath
config as we don't want it to be quoted. If quoted, it will be passed
quoted to the Clang command and Clang will fail to recognize the macro
definitions.

Fixes: 9d3f3125bd ("test/verifier: make data path verifier complexity tests standalone")
Signed-off-by: Paul Chaignon <paul@cilium.io>